Revert "Refactor Project Structure"
