+++ src/App.jsx
@@ -1,11 +1,97 @@
-import Board from "./pages/Users/_id";
+import { useColorScheme } from "@mui/material/styles";
+
+import InputLabel from "@mui/material/InputLabel";
+import MenuItem from "@mui/material/MenuItem";
+import FormControl from "@mui/material/FormControl";
+import Select from "@mui/material/Select";
+import Box from "@mui/material/Box";
+import Container from "@mui/material/Container";
+
+import LightModeIcon from "@mui/icons-material/LightMode";
+import DarkModeOutlinedIcon from "@mui/icons-material/DarkModeOutlined";
+import SettingsBrightnessIcon from "@mui/icons-material/SettingsBrightness";
+
+function ModeSelect() {
+  const { mode, setMode } = useColorScheme();
+
+  const handleChange = (event) => {
+    const selectedMode = event.target.value;
+    setMode(selectedMode);
+  };
+
+  return (
+    <FormControl sx={{ m: 1, minWidth: 120 }} size="small">
+      <InputLabel id="lable-select-dark-light-mode">Mode</InputLabel>
+      <Select
+        labelId="lable-select-dark-light-mode"
+        id="lable-select-dark-light-mode"
+        value={mode}
+        label="Age"
+        onChange={handleChange}
+      >
+        <MenuItem value="light">
+          <Box sx={{ display: "flex", alignItems: "center", gap: 1 }}>
+            <LightModeIcon fontSize="small" />
+            Light
+          </Box>
+        </MenuItem>
+        <MenuItem value="system">
+          <Box sx={{ display: "flex", alignItems: "center", gap: 1 }}>
+            <SettingsBrightnessIcon fontSize="small" />
+            System
+          </Box>
+        </MenuItem>
+        <MenuItem value="dark">
+          <Box sx={{ display: "flex", alignItems: "center", gap: 1 }}>
+            <DarkModeOutlinedIcon fontSize="small" />
+            Dark
+          </Box>
+        </MenuItem>
+      </Select>
+    </FormControl>
+  );
+}
 
 function App() {
   return (
-    <>
-      {/* React Router Dom */}
-      <Board />
-    </>
+    <Container disableGutters maxWidth={false} sx={{ height: "100vh" }}>
+      <Box
+        sx={{
+          backgroundColor: "primary.light",
+          width: "100%",
+          height: (theme) => theme.trello.appBarHeight,
+          display: "flex",
+          alignItems: "center",
+        }}
+      >
+        <ModeSelect />
+      </Box>
+
+      <Box
+        sx={{
+          backgroundColor: "primary.dark",
+          width: "100%",
+          height: (theme) => theme.trello.boardBarHeight,
+          display: "flex",
+          alignItems: "center",
+        }}
+      >
+        Board Bar
+      </Box>
+
+      <Box
+        sx={{
+          backgroundColor: "primary.main",
+          width: "100%",
+          height: (theme) =>
+            `calc(100vh - ${theme.trello.appBarHeight} - ${theme.trello.boardBarHeight})`,
+          display: "flex",
+          alignItems: "center",
+        }}
+      >
+        Board Content
+      </Box>
+    </Container>
   );
 }
