--- src/components/ModeSelect/index.jsx
@@ -1,54 +1,0 @@
-import { useColorScheme } from "@mui/material/styles";
-
-import InputLabel from "@mui/material/InputLabel";
-import MenuItem from "@mui/material/MenuItem";
-import FormControl from "@mui/material/FormControl";
-import Select from "@mui/material/Select";
-import Box from "@mui/material/Box";
-
-import LightModeIcon from "@mui/icons-material/LightMode";
-import DarkModeOutlinedIcon from "@mui/icons-material/DarkModeOutlined";
-import SettingsBrightnessIcon from "@mui/icons-material/SettingsBrightness";
-
-function ModeSelect() {
-  const { mode, setMode } = useColorScheme();
-
-  const handleChange = (event) => {
-    const selectedMode = event.target.value;
-    setMode(selectedMode);
-  };
-
-  return (
-    <FormControl sx={{ m: 1, minWidth: 120 }} size="small">
-      <InputLabel id="lable-select-dark-light-mode">Mode</InputLabel>
-      <Select
-        labelId="lable-select-dark-light-mode"
-        id="lable-select-dark-light-mode"
-        value={mode}
-        label="Age"
-        onChange={handleChange}
-      >
-        <MenuItem value="light">
-          <Box sx={{ display: "flex", alignItems: "center", gap: 1 }}>
-            <LightModeIcon fontSize="small" />
-            Light
-          </Box>
-        </MenuItem>
-        <MenuItem value="system">
-          <Box sx={{ display: "flex", alignItems: "center", gap: 1 }}>
-            <SettingsBrightnessIcon fontSize="small" />
-            System
-          </Box>
-        </MenuItem>
-        <MenuItem value="dark">
-          <Box sx={{ display: "flex", alignItems: "center", gap: 1 }}>
-            <DarkModeOutlinedIcon fontSize="small" />
-            Dark
-          </Box>
-        </MenuItem>
-      </Select>
-    </FormControl>
-  );
-}
-
-export default ModeSelect;
