--- src/pages/Boards/BoardBar/index.jsx
@@ -1,19 +1,0 @@
-import Box from "@mui/material/Box";
-
-function BoardBar() {
-  return (
-    <Box
-      sx={{
-        backgroundColor: "primary.dark",
-        width: "100%",
-        height: (theme) => theme.trello.boardBarHeight,
-        display: "flex",
-        alignItems: "center",
-      }}
-    >
-      Board Bar
-    </Box>
-  );
-}
-
-export default BoardBar;
