--- src/pages/Boards/BoardContent/index.jsx
@@ -1,20 +1,0 @@
-import Box from "@mui/material/Box";
-
-function BoardContent() {
-  return (
-    <Box
-      sx={{
-        backgroundColor: "primary.main",
-        width: "100%",
-        height: (theme) =>
-          `calc(100vh - ${theme.trello.appBarHeight} - ${theme.trello.boardBarHeight})`,
-        display: "flex",
-        alignItems: "center",
-      }}
-    >
-      Board Content
-    </Box>
-  );
-}
-
-export default BoardContent;
