--- src/pages/Users/_id.jsx
@@ -1,1 +1,0 @@
-// User Detail
