--- src/pages/Users/index.jsx
@@ -1,1 +1,0 @@
-//Users list
