--- src/utils/constants.js
@@ -1,1 +1,0 @@
-export const API_ROOT = "https://localhost:8017";
